requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose
cryptography
passlib
python-dotenv
pydantic
fastapi-cli
pydantic[email]
python-multipart